// frontend/src/typechain-types/contracts/PropertyManager.ts
export type PropertyStructOutput = [
  BigNumber,
  string,
  string,
  BigNumber,
  BigNumber,
  string[],
  string,
  number,
  boolean,
  string
] & {
  id: BigNumber;
  name: string;
  owner: string;
  price: BigNumber;
  funds: BigNumber;
  images: string[];
  fundStorage: string;
  status: number;
  verified: boolean;
  location: string;
};